+++ packages/adhere-ui-antdformitem/src/AntFormItemNormalize/index.tsx
@@ -65,8 +65,9 @@ import {
   Typography as AntTypography,
   Upload as AntUpload,
   Watermark as AntWatermark,
+  ButtonProps,
 } from 'antd';
-import React from 'react';
+import React, { FC, useState } from 'react';
 
 import FlexLayout from '@baifendian/adhere-ui-flexlayout';
 import Resource from '@baifendian/adhere-util-resource';
@@ -84,7 +85,6 @@ const { useScrollLayout } = FlexLayout;
  */
 function createFactory(Component, defaultProps): any {
   function fn(_props) {
-    // eslint-disable-next-line react-hooks/rules-of-hooks
     const { getEl } = useScrollLayout();
 
     const props = { ...defaultProps, ..._props };
@@ -96,8 +96,6 @@ function createFactory(Component, defaultProps): any {
     }
 
     const { children, ...reset } = props;
-
-    console.log('reset', Component, reset);
 
     return <Component {...reset}>{children}</Component>;
   }
@@ -209,6 +207,43 @@ export const InputNumberDecimal2 = createFactory(AntInputNumber, {
 export const InputNumberInteger = createFactory(AntInputNumber, {
   precision: 0,
 });
+
+/**
+ * SubmitButton
+ * @description 提交按钮
+ */
+export const SubmitButton = createFactory(
+  (function () {
+    const Component: FC<ButtonProps> = (props) => {
+      const [loading, setLoading] = useState<boolean>(false);
+
+      return (
+        <Button
+          loading={loading}
+          {...props}
+          onClick={(e) => {
+            if (!props.onClick) return;
+
+            if (loading) return;
+
+            setLoading(true);
+
+            props
+              .onClick(e)
+              // @ts-ignore
+              ?.then?.(() => setLoading(false))
+              ?.catch?.(() => setLoading(false));
+          }}
+        >
+          {props.children}
+        </Button>
+      );
+    };
+
+    return Component;
+  })(),
+  {},
+);
 
 export const Affix = createFactory(AntAffix, {});
 export const Alert = createFactory(AntAlert, {});